refactor grid related directories
