--- frontend/src/components/Collections/components/CollectionsGrid/components/common/style.ts
@@ -1,13 +1,0 @@
-import styled, { css } from "styled-components";
-
-export const textClippingCSS = css`
-  white-space: nowrap;
-`;
-
-export const detailsColWidthCSS = css`
-  width: calc(1 / 8 * 100%);
-`;
-
-export const StyledCell = styled.td`
-  vertical-align: middle !important;
-`;
