+++ frontend/src/components/Collections/components/Grid/components/CollectionsGrid/index.tsx
@@ -1,13 +1,13 @@
 import React, { FC } from "react";
 import { ACCESS_TYPE } from "src/common/entities";
 import { CollectionResponse } from "src/common/queries/collections";
-import CollectionRow from "./components/CollectionRow";
 import {
   CollectionHeaderCell,
   LeftAlignedHeaderCell,
   RightAlignedHeaderCell,
   StyledCollectionsGrid,
-} from "./style";
+} from "../../common/style";
+import CollectionRow from "../Row/CollectionRow";
 
 interface Props {
   collections: CollectionResponse[];
