+++ frontend/src/components/Collections/components/Grid/components/Popover/style.ts
@@ -1,6 +1,6 @@
 import { GRAY, PT_GRID_SIZE_PX } from "src/components/common/theme";
 import styled from "styled-components";
-import { textClippingCSS } from "../../../common/style";
+import { textClippingCSS } from "../../common/style";
 
 export const FieldValues = styled.div`
   ${textClippingCSS}
